requirements.txt
streamlit
feedparser
pandas
python-dotenv
openai
plotly
